[ALPS03971604] PTPOD: Enable upower spower config
[Detail] PTPOD: Enable upower spower config and change cap type

MTK-Commit-Id: 3aad6621ea24d2278269b02be7af372c8f4c307e

CR-Id: ALPS03971604
Feature: PTP-OD (Performance Thermal Power-Over Drive)
(cherry-pick from 72fee38440843fa14b9e762a18a44fd151b359b3)

+++ drivers/misc/mediatek/base/power/include/upower_v2/mtk_unified_power.h
@@ -60,7 +60,7 @@ extern "C" {
 /* but compiler will align to 40 bytes for computing more faster */
 /* if a table has 16 opps --> 40*16= 640 bytes*/
 struct upower_tbl_row {
-	unsigned long long cap;
+	unsigned long cap;
 	unsigned int volt; /* 10uv */
 	unsigned int dyn_pwr; /* uw */
 	unsigned int lkg_pwr[NR_UPOWER_DEGREE]; /* uw */
